refactor: move logic out of layout component
